+++ cmd/bd/show_display.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/steveyegge/beads/internal/types"
 	"github.com/steveyegge/beads/internal/ui"
+	"github.com/steveyegge/beads/internal/uimd"
 )
 
 // displayShowIssue displays a single issue (reusable for watch mode).
@@ -102,16 +103,16 @@ func displayShowIssueReturn(ctx context.Context, issueID string) *types.Issue {
 
 	// Content sections (matches standard bd show order)
 	if issue.Description != "" {
-		fmt.Printf("\n%s\n%s\n", ui.RenderBold("DESCRIPTION"), ui.RenderMarkdown(issue.Description))
+		fmt.Printf("\n%s\n%s\n", ui.RenderBold("DESCRIPTION"), uimd.RenderMarkdown(issue.Description))
 	}
 	if issue.Design != "" {
-		fmt.Printf("\n%s\n%s\n", ui.RenderBold("DESIGN"), ui.RenderMarkdown(issue.Design))
+		fmt.Printf("\n%s\n%s\n", ui.RenderBold("DESIGN"), uimd.RenderMarkdown(issue.Design))
 	}
 	if issue.Notes != "" {
-		fmt.Printf("\n%s\n%s\n", ui.RenderBold("NOTES"), ui.RenderMarkdown(issue.Notes))
+		fmt.Printf("\n%s\n%s\n", ui.RenderBold("NOTES"), uimd.RenderMarkdown(issue.Notes))
 	}
 	if issue.AcceptanceCriteria != "" {
-		fmt.Printf("\n%s\n%s\n", ui.RenderBold("ACCEPTANCE CRITERIA"), ui.RenderMarkdown(issue.AcceptanceCriteria))
+		fmt.Printf("\n%s\n%s\n", ui.RenderBold("ACCEPTANCE CRITERIA"), uimd.RenderMarkdown(issue.AcceptanceCriteria))
 	}
 
 	// Labels
@@ -212,7 +213,7 @@ func displayShowIssueReturn(ctx context.Context, issueID string) *types.Issue {
 		fmt.Printf("\n%s\n", ui.RenderBold("COMMENTS"))
 		for _, comment := range comments {
 			fmt.Printf("  %s %s\n", ui.RenderMuted(comment.CreatedAt.UTC().Format("2006-01-02 15:04")), comment.Author)
-			rendered := ui.RenderMarkdown(comment.Text)
+			rendered := uimd.RenderMarkdown(comment.Text)
 			for _, line := range strings.Split(strings.TrimRight(rendered, "\n"), "\n") {
 				fmt.Printf("    %s\n", line)
 			}
